fix: 마이페이지에서 유저 정보를 서버사이드에서 가져오도록 수정
