+++ package.json
@@ -18,11 +18,11 @@
     ]
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.59.0",
-    "@tanstack/react-query-devtools": "^5.59.0",
     "@mdx-js/loader": "^3.1.0",
     "@mdx-js/react": "^3.1.0",
     "@next/mdx": "^15.0.2",
+    "@tanstack/react-query": "^5.59.0",
+    "@tanstack/react-query-devtools": "^5.59.0",
     "@types/mdx": "^2.0.13",
     "axios": "^1.7.7",
     "dayjs": "^1.11.13",
